backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pyodbc==5.0.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
slowapi==0.1.9
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httpx==0.25.2

# ML Libraries (compatible with Python 3.14)
torch>=2.9.0
torchvision>=0.20.0
transformers==4.36.0
xgboost>=2.0.3
scikit-learn>=1.5.0
pandas>=2.1.4
numpy>=1.26.2
pillow>=10.1.0
joblib>=1.3.2
